+++ functions/package.json
@@ -19,7 +19,7 @@
     "cookie-parser": "^1.4.6",
     "dotenv": "^16.0.1",
     "firebase-admin": "^10.0.2",
-    "firebase-functions": "^3.24.0"
+    "firebase-functions": "^4.0.2"
   },
   "devDependencies": {
     "@types/cookie-parser": "^1.4.3",
