+++ functions/src/user/user.controller.ts
@@ -1,15 +1,15 @@
-import {cloudFunction, axiosInstance} from "../index";
-import {NextTracksBody, UserResponse} from "./user.interface";
-import {isAuthenticated} from "../common/common.error";
-import {TokenBody} from "../common/common.interface";
+import { cloudFunction, axiosInstance } from '../index';
+import { ITopResponse, NextTracksBody, UserResponse } from './user.interface';
+import { isAuthenticated } from '../common/common.error';
+import { TokenBody } from '../common/common.interface';
 
-const URL = "https://api.spotify.com/v1/me";
+const URL = 'https://api.spotify.com/v1/me';
 
 export const getSpotifyProfile = cloudFunction.onCall(async (data: TokenBody, context) => {
   isAuthenticated(context);
 
   const results = await axiosInstance(URL, {
-    headers: {Authorization: `Bearer ${data.token}`},
+    headers: { Authorization: `Bearer ${data.token}` },
   });
 
   const profile: UserResponse = results.data;
@@ -21,15 +21,34 @@ export const getSpotifyProfile = cloudFunction.onCall(async (data: TokenBody, co
   };
 });
 
+export const getTopItems = cloudFunction.onCall(async (data: TokenBody, context) => {
+  isAuthenticated(context);
+
+  const type = 'artists';
+
+  const results = await axiosInstance(`${URL}/top/${type}?time_range=short_term&limit=3`, {
+    headers: { Authorization: `Bearer ${data.token}` },
+  });
+
+  const topItems: ITopResponse = results.data;
+
+  const itemToResponse = topItems.items.map(item => ({
+    artowrk: item.images[1].url,
+    artist: item.name,
+  }));
+
+  return itemToResponse;
+});
+
 export const getInitialSavedTracks = cloudFunction.onCall(async (data: TokenBody, context) => {
   isAuthenticated(context);
 
-  const market = "market=ES";
-  const limit = "limit=50";
-  const offset = "offset=0";
+  const market = 'market=ES';
+  const limit = 'limit=50';
+  const offset = 'offset=0';
 
   const results = await axiosInstance(`${URL}/tracks?${market}&${limit}&${offset}`, {
-    headers: {Authorization: `Bearer ${data.token}`},
+    headers: { Authorization: `Bearer ${data.token}` },
   });
   return results.data;
 });
@@ -38,7 +57,7 @@ export const getNextSavedTracks = cloudFunction.onCall(async (data: NextTracksBo
   isAuthenticated(context);
 
   const results = await axiosInstance(data.url, {
-    headers: {Authorization: `Bearer ${data.token}`},
+    headers: { Authorization: `Bearer ${data.token}` },
   });
 
   return results.data;
