--- src/Components/atoms/Button/Button.scss
@@ -1,19 +1,0 @@
-@use '../../../styles/colors';
-
-.button {
-  width: 80vw;
-  height: 50px;
-
-  border-radius: 5px;
-  border: 1px solid colors.$border;
-
-  color: black;
-  font-weight: 600;
-  font-size: 14px;
-  line-height: 27px;
-  letter-spacing: 0.1em;
-
-  &:hover {
-    cursor: pointer;
-  }
-}
