--- src/Components/atoms/PlaybackButton/PlaybackButton.scss
@@ -1,16 +1,0 @@
-@use '../../../styles/colors';
-
-.playback-button {
-  width: 100px;
-
-  svg {
-    width: 30px;
-    height: 30px;
-  }
-
-  &:hover {
-    path {
-      stroke: colors.$grey-1;
-    }
-  }
-}
