+++ src/Components/molecules/Player/Player.tsx
@@ -2,11 +2,8 @@ import './Player.scss';
 import * as React from 'react';
 import { useAppSelector } from '../../../store/hooks';
 import useSpotifySDK from '../../../hooks/useSpotifySDK';
-import { PlaybackButton } from '../../atoms';
-import {
-  isActiveSelector,
-  playBackInfoSelector,
-} from '../../../store/playback/playback.slice';
+import { Body, PlaybackButton } from '../../atoms';
+import { isActiveSelector, playBackInfoSelector } from '../../../store/playback/playback.slice';
 
 const Player = () => {
   const player = useSpotifySDK();
@@ -31,8 +28,8 @@ export const PlayBackInfo = () => {
     <div className="player__track">
       <img src={image} alt="album cover" className="player__artwork" />
       <div className="player__info">
-        <p className="player__info--name">{name}</p>
-        <p className="player__info--artist">{artist}</p>
+        <Body className="player__info--name">{name}</Body>
+        <Body>{artist}</Body>
       </div>
     </div>
   );
