--- src/Pages/home/Home.tsx
@@ -1,8 +1,0 @@
-import * as React from 'react';
-import Dashboard from './Dashboard';
-
-const Home = () => {
-  return <Dashboard />;
-};
-
-export default Home;
