--- src/Pages/tracks/Tracks.scss
@@ -1,15 +1,0 @@
-@use '../../styles/breakpoints';
-
-.tracks-view {
-  display: flex;
-  min-height: 100vh;
-  min-height: -webkit-fill-available;
-  overflow: hidden;
-
-  @media screen and (min-width: breakpoints.$phone) {
-    display: flex;
-    justify-content: space-evenly;
-    align-items: center;
-    height: 90vh;
-  }
-}
